formatting

+++ src/Illuminate/Queue/Worker.php
@@ -396,9 +396,7 @@ class Worker
             // so it is not lost entirely. This'll let the job be retried at a later time by
             // another listener (or this same one). We will re-throw this exception after.
             if (! $job->isDeleted() && ! $job->isReleased() && ! $job->hasFailed()) {
-                $job->release(
-                    $this->calculateBackoff($job, $options)
-                );
+                $job->release($this->calculateBackoff($job, $options));
             }
         }
 
@@ -504,13 +502,14 @@ class Worker
      */
     protected function calculateBackoff($job, WorkerOptions $options)
     {
-        $backoff = explode(',',
+        $backoff = explode(
+            ',',
             method_exists($job, 'backoff') && ! is_null($job->backoff())
                         ? $job->backoff()
                         : $options->backoff
         );
 
-        return $backoff[$job->attempts() - 1] ?? last($backoff);
+        return (int) ($backoff[$job->attempts() - 1] ?? last($backoff));
     }
 
     /**
